restructured mtraining
