+++ mtraining/experiments/train_attn_configs/qwen_mf_dr_stripe.yaml
