+++ mtraining/experiments/train_attn_configs/qwen_mf_zigzag.yaml
