+++ mtraining/model_configs/phi3/lc_config_mini/configuration_phi3.py
