--- mtraining/models/sparse_ops/.gitignore
@@ -1,7 +1,0 @@
-__pycache__/
-*.egg-info/
-build/
-*.egg
-configs/
-minference_attn.py
-minference_sparse_index.py
