--- mtraining/models/sparse_ops/mtraining_sparse_ops/__init__.py
@@ -1,2 +1,0 @@
-from .minference_config import get_minference_config
-from .minference_attn import minference_flash_attn_func
