+++ mtraining/setup.sh
@@ -8,7 +8,7 @@ if command -v nvidia-smi
 then
     # assume base image: amlt-sing/acpt-torch2.3.1-py3.10-cuda12.1-ubuntu22.04
     $PIP install ninja cmake wheel pybind11
-    $PIP install --no-cache-dir torch==2.3.1 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
+    $PIP install --no-cache-dir torch==2.3.1 --index-url https://download.pytorch.org/whl/cu121
     $PIP install git+https://github.com/Dao-AILab/flash-attention.git@v2.7.4.post1
     $PIP install -r "${BASE_DIR}/requirements.txt"
     $PIP install git+https://github.com/microsoft/nnscaler.git@2368540417bc3b77b7e714d3f1a0de8a51bb66e8
@@ -17,7 +17,7 @@ then
 elif command -v rocm-smi
 then
     $PIP install ninja cmake wheel pybind11
-    $PIP install --no-cache-dir --pre torch==2.3.1+rocm6.0 torchvision torchaudio --index-url https://download.pytorch.org/whl/rocm6.0
+    $PIP install --no-cache-dir --pre torch==2.3.1+rocm6.0 --index-url https://download.pytorch.org/whl/rocm6.0
     $PIP install git+https://github.com/OpenAI/triton.git@e192dba#subdirectory=python
     $PIP install git+https://github.com/Dao-AILab/flash-attention.git@v2.7.4.post1
     $PIP install -r "${BASE_DIR}/requirements.txt"
@@ -32,3 +32,4 @@ NNSCALER_HOME=$(python -c "import nnscaler; print(nnscaler.__path__[0])")
 echo "export NNSCALER_HOME=${NNSCALER_HOME}" >> ~/.profile
 echo "export PYTHONPATH=${NNSCALER_HOME}:\${PYTHONPATH}" >> ~/.profile
 source ~/.profile
+pip install -e .
